Launch the shinybrms Shiny app

This function launches the shinybrms Shiny app.

launch_shinybrms(...)

Arguments

...
Arguments passed to shiny::runApp(). For example, in order to force launching the Shiny app in the system's default web browser (even if using RStudio), set argument launch.browser to TRUE.

Value

TRUE (invisible).

Note

The following global options are introduced by shinybrms:

shinybrms.prog_browser: The value for the global option browser to be used for opening the Stan progress file (an HTML file). Defaults to option browser (see base::options()). Note that RStudio's default for option browser cannot be used for the Stan progress file. Therefore, shinybrms automatically redefines RStudio's default for option browser to fix this issue.

shinybrms.shinystan_browser: The value for the global option browser to be used for opening the shinystan app. Defaults to option browser (see base::options()). Note that RStudio's default for option browser cannot be used for opening the shinystan app from within shinybrms. Therefore, shinybrms automatically redefines RStudio's default for option browser to fix this issue.

The only other global option used by shinybrms is mc.cores (see base::options()): If option mc.cores is set, then the advanced option "Cores" (in the shinybrms app) defaults to the value of this option. If option mc.cores is not set, then the advanced option "Cores" defaults to the value returned by parallel::detectCores().

Note that shinystan's global option shinystan.rstudio (see shinystan::launch_shinystan()) may not be used here as shinystan is called from an external R process where RStudio's browsers are not available.

Examples

if(interactive()){
  launch_shinybrms(launch.browser = TRUE)
}